backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
ciso8601==2.2.0
python-rapidjson==1.4
tortoise-orm[asyncpg]==0.17.8
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
minio==7.1.1